+++ source/new_ppo.py
@@ -21,102 +21,8 @@ from collections import namedtuple
 
 from base import Agent, Model
 
-# Other parameters
-frame_repeat = 12
-resolution = (30, 45)
-
 # A named tuple to save information about steps that have been performed by the net and what came back from the net
 Experience = namedtuple('Experience', ('state', 'action', 'action_log_prob', 'value', 'reward', 'mask'))
-
-# Uses GPU if available
-if torch.cuda.is_available():
-    DEVICE = torch.device('cuda')
-    torch.backends.cudnn.benchmark = True
-else:
-    DEVICE = torch.device('cpu')
-    
-def preprocess(img):
-    """Down samples image to resolution"""
-    img = skimage.transform.resize(img, resolution)
-    img = img.astype(np.float32)
-    img = np.expand_dims(img, axis=0)
-
-    return img
-
-def test(game, agent):
-    """Runs a test_episodes_per_epoch episodes and prints the result"""
-    print("\nTesting...")
-    test_scores = []
-    for test_episode in trange(test_episodes_per_epoch, leave=False):
-        game.new_episode()
-        while not game.is_episode_finished():
-            state = preprocess(game.get_state().screen_buffer)
-            best_action_index = agent.get_action(state)
-
-            game.make_action(actions[best_action_index], frame_repeat)
-        r = game.get_total_reward()
-        test_scores.append(r)
-
-    test_scores = np.array(test_scores)
-    print("Results: mean: %.1f +/- %.1f," % (
-        test_scores.mean(), test_scores.std()), "min: %.1f" % test_scores.min(),
-          "max: %.1f" % test_scores.max())
-
-#TODO: check this.
-def run(game, agent, actions, num_epochs, frame_repeat, steps_per_epoch=2000):
-    """
-    Run num epochs of training episodes.
-    Skip frame_repeat number of frames after each action.
-    """
-
-    start_time = time()
-
-    for epoch in range(num_epochs):
-        game.new_episode()
-        train_scores = []
-        global_step = 0
-        print("\nEpoch #" + str(epoch + 1))
-
-        for _ in trange(steps_per_epoch, leave=False):
-            state = preprocess(game.get_state().screen_buffer)
-            action = agent.get_action(state)
-            reward = game.make_action(actions[action], frame_repeat)
-            done = game.is_episode_finished()
-
-            if not done:
-                next_state = preprocess(game.get_state().screen_buffer)
-            else:
-                next_state = np.zeros((1, 90, 135)).astype(np.float32) # 30x45 = 30x45
-
-            agent.append_memory(state, action, reward, next_state, done)
-
-            if global_step > agent.batch_size:
-                agent.train()
-
-            if done:
-                train_scores.append(game.get_total_reward())
-                game.new_episode()
-
-            global_step += 1
-
-        agent.update_target_net()
-        overall_train_score.append(train_scores)
-        with open(text_file,"w") as textfile:
-            textfile.write(str(overall_train_score))
-        train_scores = np.array(train_scores)
-        
-
-        print("Results: mean: %.1f +/- %.1f," % (train_scores.mean(), train_scores.std()),
-              "min: %.1f," % train_scores.min(), "max: %.1f," % train_scores.max())
-
-        test(game, agent)
-        if save_model:
-            print("Saving the network weights to:", model_savefile)
-            torch.save(agent.q_net, model_savefile)
-        print("Total elapsed time: %.2f minutes" % ((time() - start_time) / 60.0))
-
-    game.close()
-    return agent, game
    
 class ActorCritc(Model):
     
@@ -253,3 +159,5 @@ class PPOAgent(Agent):
         adv = np.array(returns) - values[:-1]
         return returns, (adv - np.mean(adv)) / (np.std(adv) + 1e-10)
     
+    def train():
+        pass
